完成游戏主要逻辑编写

+++ logicserver/src/main/java/com/linkflywind/gameserver/logicserver/protocolData/A1011Response.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class A1011Response {
-    private int zhuoMain;
+    private double zhuoMain;
     YingSanZhangPlayer[] players;
 }
